+++ migrations/003_guildconfig.py
@@ -1,0 +1,59 @@
+"""Peewee migrations -- 003_guildconfig.py.
+
+Some examples (model - class or model name)::
+
+    > Model = migrator.orm['table_name']            # Return model in current state by name
+    > Model = migrator.ModelClass                   # Return model in current state by name
+
+    > migrator.sql(sql)                             # Run custom SQL
+    > migrator.python(func, *args, **kwargs)        # Run python code
+    > migrator.create_model(Model)                  # Create a model (could be used as decorator)
+    > migrator.remove_model(model, cascade=True)    # Remove a model
+    > migrator.add_fields(model, **fields)          # Add fields to a model
+    > migrator.change_fields(model, **fields)       # Change fields
+    > migrator.remove_fields(model, *field_names, cascade=True)
+    > migrator.rename_field(model, old_field_name, new_field_name)
+    > migrator.rename_table(model, new_table_name)
+    > migrator.add_index(model, *col_names, unique=False)
+    > migrator.drop_index(model, *col_names)
+    > migrator.add_not_null(model, *field_names)
+    > migrator.drop_not_null(model, *field_names)
+    > migrator.add_default(model, field_name, default)
+
+"""
+
+import peewee as pw
+from peewee_migrate import Migrator
+from decimal import ROUND_HALF_EVEN
+
+try:
+    import playhouse.postgres_ext as pw_pext
+except ImportError:
+    pass
+
+SQL = pw.SQL
+
+
+def migrate(migrator: Migrator, database: pw.Database, *, fake=False):
+    """Write your migrations here."""
+    
+    migrator.remove_fields('bibbia', 'id')
+
+    migrator.add_fields(
+        'guildconfig',
+
+        daytime_end=pw.IntegerField(constraints=[SQL("DEFAULT 1380")], default=1380, null=True),
+        cctv_channel_id=pw.BigIntegerField(null=True),
+        main_role_id=pw.BigIntegerField(null=True),
+        daytime_start=pw.IntegerField(constraints=[SQL("DEFAULT 420")], default=420, null=True))
+
+
+def rollback(migrator: Migrator, database: pw.Database, *, fake=False):
+    """Write your rollback migrations here."""
+    
+    migrator.remove_fields('guildconfig', 'daytime_end', 'cctv_channel_id', 'main_role_id', 'daytime_start')
+
+    migrator.add_fields(
+        'bibbia',
+
+        id=pw.AutoField())