+++ mukurobot/dbutils.py
@@ -2,7 +2,7 @@ from playhouse.shortcuts import ReconnectMixin
 from peewee import MySQLDatabase
 
 class ReconnectMysqlDatabase(ReconnectMixin, MySQLDatabase):
-    pass
+    _reconnect_errors = ReconnectMixin.reconnect_errors
 
 def connect_reconnect(db):
     if db.__class__ == MySQLDatabase: