+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 0.2
 
++ Enhanced guild configs.
 + Changed calls to `print()` to logging calls.
-+ Added `utils.py` module.
++ Added timers.
++ Added `utils.py`, `dbutils.py`, `dsutils.py` and `timers.py` modules.
 + Added `/bibbia` and `/lore` commands.
 
 ## 0.1.1